+++ OwnTracks/OwnTracks/Connection.m
@@ -322,7 +322,8 @@ static const DDLogLevel ddLogLevel = DDLogLevelError;
                    @(MQTTSessionEventConnectionRefused): @"connection refused",
                    @(MQTTSessionEventConnectionClosed): @"connection closed",
                    @(MQTTSessionEventConnectionError): @"connection error",
-                   @(MQTTSessionEventProtocolError): @"protocoll error"
+                   @(MQTTSessionEventProtocolError): @"protocol error"
+                   @(MQTTSessionEventConnectionClosedByBroker): @"connection closed by broker"
                    }[@(eventCode)],
                  (long)eventCode, error);